+++ crates/biome_js_analyze/tests/specs/nursery/useNumberToFixedDigitsArgument/valid.js
@@ -1,0 +1,13 @@
+number.toFixed(0);
+number.toFixed(...[]);
+number.toFixed(2);
+number.toFixed(1, 2, 3);
+number[toFixed]();
+number["toFixed"]();
+number?.toFixed();
+number.toFixed?.();
+number.notToFixed();
+
+// `callee` is a `NewExpression`
+new BigNumber(1).toFixed();
+new Number(1).toFixed();