+++ packages/@biomejs/backend-jsonrpc/src/workspace.ts
@@ -1049,6 +1049,10 @@ export interface Nursery {
 	 */
 	useImportRestrictions?: RuleConfiguration_for_Null;
 	/**
+	 * Enforce using the digits argument with Number#toFixed().
+	 */
+	useNumberToFixedDigitsArgument?: RuleConfiguration_for_Null;
+	/**
 	 * Enforce the sorting of CSS utility classes.
 	 */
 	useSortedClasses?: RuleConfiguration_for_UtilityClassSortingOptions;
@@ -2058,6 +2062,7 @@ export type Category =
 	| "lint/nursery/useFocusableInteractive"
 	| "lint/nursery/useGenericFontNames"
 	| "lint/nursery/useImportRestrictions"
+	| "lint/nursery/useNumberToFixedDigitsArgument"
 	| "lint/nursery/useSortedClasses"
 	| "lint/nursery/useThrowNewError"
 	| "lint/performance/noAccumulatingSpread"